Virus explosions
This should be the last update for virus explosions.
These were compared directly to the vanilla explosions of multiple party servers (recorded in spectate). After a couple hours of research and tweaking the multipliers I have found that there are 3 different types of explosions. I have adjusted the multipliers for each of them to as vanilla as possible, each explosion is selected randomly when you hit a virus. Should also make popsplit rates the same as vanilla. (turns out they don't work as often as I thought they did)